refactor: rename file to have same name as class
